split blockHarvester
