+++ lib/crawlers/subscribeToNewBlocks.js
@@ -1,7 +1,7 @@
 const { shortHash } = require('../utils.js');
 
 module.exports = {
-  subscribeToNewBlocks: async function subscribeToNewBlocks(api, pool) {
+  subscribeToNewBlocks: async function (api, pool) {
     // Subscribe to new blocks
     await api.rpc.chain.subscribeNewHeads(async (header) => {
